try threaded tests
